added "readline()" builtin function 	fixed some bugs with widgets (buttons I believe) being deleted 		out from under us 	added 'newcmap=F' parameter to "frame()" 	also added 'mincolors=2', 'maxcolors=100', and 'cmapfail=F' to 		pgplot(), but TkPgplot.cc and TkPgplot.h aren't under 		CVS... they are in the AIPS++ system...

+++ glish/Value.cc
@@ -1014,7 +1014,7 @@ dcomplex* Value::CoerceToDcomplexArray( int& is_copy, int size,
 
 	switch ( Type() )
 		{
-		COERCE_COMPLEX_ACTIONS(float,"CoerceToDcomplexArray()")
+		COERCE_COMPLEX_ACTIONS(double,"CoerceToDcomplexArray()")
 		default:
 			error->Report(
 			    "bad type in Value::CoerceToDcomplexArray()" );
